@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20130604183834) do
+ActiveRecord::Schema.define(version: 20150916194447) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -30,7 +30,7 @@ ActiveRecord::Schema.define(version: 20130604183834) do
     t.text     "body"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.text     "tags",       array: true
+    t.text     "tags",       default: "--- []\n"
     t.boolean  "draft"
   end
 
